chore: Move `DatabaseConfig` into `cda-config`.
