+++ cda-config/src/datatypes/mod.rs
@@ -11,10 +11,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+pub use database::*;
 pub use database_naming_convention::*;
 pub use doip::*;
 pub use functional_description::*;
 
+mod database;
 mod database_naming_convention;
 mod doip;
 mod functional_description;
